% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_historical_track.R
\name{calculate_historical_track}
\alias{calculate_historical_track}
\title{functions to give out the hitorical record according to the returns calculated from strategy chosen}
\usage{
calculate_historical_track(dataframe, portfolio, weighting)
}
\arguments{
\item{dataframe, }{stocks inputs..}

\item{portfolio, }{portfolio chosen from the stock calculation as a return base}

\item{weighting, }{determines the weight of period as a sliding window}
}
\value{
return binded result from historical calculation of stocks based on chosen investment target from portfolio calculation
}
\description{
functions to give out the hitorical record according to the returns calculated from strategy chosen
}
\details{
code to calculate returuns bases on a portfolio determined via the function determine_portfolio
}
\examples{
calculate_historical_track(dataframe, portfolio, weighting)
}
