% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_upside_downside.R
\name{compute_upside_downside}
\alias{compute_upside_downside}
\title{draw upside trend, downside trend regime of chosen stocks, compared to bench mark stock}
\usage{
compute_upside_downside(stock, benchmark)
}
\arguments{
\item{stock}{single stock information}

\item{benchmark}{bench mark stock price user choose}
}
\value{
plot of computed upside, downside trend of stocks user choose
}
\description{
draw upside trend, downside trend regime of chosen stocks, compared to bench mark stock
}
\details{
this function will take the historical prices of a stock and its benchmark and will do following computations.
consider returns only positive and ignore all negative returns: compute an historical track as negative days didn't happen
similar function for days as if positive days didn't happen
this metric allows to see how much you do better during upside markets and downside markets separately
}
\examples{
compute_upside_downside(stock, benchmark_stock)
}
