% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_portfolio.R
\name{determine_portfolio}
\alias{determine_portfolio}
\title{functions to choose best stock to invest based on strategy}
\usage{
determine_portfolio(output_variable, num_companies)
}
\arguments{
\item{output_variable, }{stocks that will be invested}

\item{num_companies}{chosen numbers what user want to have as a potential investment stock}
}
\value{
return the determined stock features based on strategy chosen
}
\description{
functions to choose best stock to invest based on strategy
}
\details{
for each date or quarter, we sort the data and select top , then give out the determined best top(chosen num_compnies) stock to be invested
}
\examples{
determine_portfolio(output_variable, num_companies)
}
