% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cluster_analysis.R
\name{draw_cluster_analysis}
\alias{draw_cluster_analysis}
\title{draw k-mean cluster analysis result in histogram table}
\usage{
draw_cluster_analysis(df_stocks, num_center, num_start)
}
\arguments{
\item{df_stocks}{bunch of stocks stacked which user want to analyze}

\item{num_center}{number of centers of clusters user might choose. There are no statisticaly proven best number. But You can use elbow-method to choose best param}

\item{num_start}{number of random initialized starting point as a initial poitn of model training}
}
\value{
plot of cluster result from k-means
}
\description{
draw k-mean cluster analysis result in histogram table
}
\details{
This compute the k-means algorithm with user designed parameter. Then it will draw the histogram result. This can be used to analyze the market trend, similiarity, risk parity, and regime of the market movement for each stocks inputted
}
\examples{
draw_cluster_analysis(df_stocks, num_center, num_start)
}
