% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_comparison_result_plot.R
\name{draw_comparison_result_plot}
\alias{draw_comparison_result_plot}
\title{draw multiple plots based on rolling regression prediction}
\usage{
draw_comparison_result_plot(input, rolled_input)
}
\arguments{
\item{rolled_input}{calculated from function execute_roll_linear}

\item{target}{single stock}
}
\value{
draws plots
}
\description{
draw multiple plots based on rolling regression prediction
}
\details{
draws the plot of showing all upper,lower,middle boundary prediction for each given period. It will compute the all necessary tool to draw the rolling sliding window based linear regression, it will draw the line for each corresponding purpose, as commented
}
\examples{
prediction_matrix= draw_comparison_result_plot(stock, rolled_input)
}
