% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pca_var_plot.R
\name{draw_pca_var_plot}
\alias{draw_pca_var_plot}
\title{# draws result of the variance analysis from pca}
\usage{
draw_pca_var_plot(stock_bunch)
}
\arguments{
\item{stock_bunch}{stacked or one matrix of multiple stock informations with only one single price information(not all OHLC)}
}
\value{
pca scatter plot
}
\description{
# draws result of the variance analysis from pca
}
\details{
this gives result by scatter plot result of PCA analysis information for current market in given period. Shows the relationship of each period for each stock
}
\examples{
draw_pca_var_plot(stock_bunch)
}
