% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_stock_analysis.R
\name{draw_stock_analysis}
\alias{draw_stock_analysis}
\title{draw bollinger band, candle chart, trick char, volume traded chart}
\usage{
draw_stock_analysis(single_stock, ticker_name, sdv, n1, n2, n3,
  last_period_months)
}
\arguments{
\item{single_stock}{single stock informations with only one single price information(all OHLC)}

\item{ticker_name}{ticker name which will be used for the analysis}

\item{sdv}{standard deviation used to make and interprete BB chart}

\item{n1}{numbers for the BBands}

\item{n2}{moving average term fomr the first line of SMA}

\item{n3}{moving average term fomr the second line of SMA}

\item{months}{Selected user chosen months to calculate BB plots}
}
\value{
Summary of bollinger band, candle chart, trick char, volume traded chart analysis result
}
\description{
draw bollinger band, candle chart, trick char, volume traded chart
}
\details{
This gives all summary of useful plots listed above from analysis information result for the single stock. to interprete bollinger bands graph: use statsitcal terms.
}
\examples{
draw_stock_analysis(single_stock, ticker_name, sdv, n1,n2,n3,last_period_months) When user use sd=2, 95\% of the data will fall in between the bollinger bands upper and lower bands.User can use it as a target and analyze the market. Lower bollinger means market will be soon moving into bull market.this will show the moving average graph also. this shows overbought and oversold situations. generally, >60 means overbought, <-50 means oversold.here it will automatically calculate the adjusted OHLC
}
