% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_var_explanation.R
\name{draw_var_explanation}
\alias{draw_var_explanation}
\title{# draws result of the variance analysis}
\usage{
draw_var_explanation(stock_bunch)
}
\arguments{
\item{stock_bunch}{stacked or one matrix of multiple stock informations with only one single price information(not all OHLC)}
}
\value{
histogram of varaince analysis
}
\description{
# draws result of the variance analysis
}
\details{
this gives result by histogram table of PCA analysis information for current market in given period
}
\examples{
draw_var_explanation(stock_bunch)
}
