% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_roll_linear.R
\name{execute_roll_linear}
\alias{execute_roll_linear}
\title{predicted value from rolling based regression}
\usage{
execute_roll_linear(a, window_size)
}
\arguments{
\item{target}{single stock}

\item{rolling}{window size}
}
\value{
matrix of predicted value,and uppder,lower boundary
}
\description{
predicted value from rolling based regression
}
\details{
this function will return the rolled_input return value which contains 1) upper boundary preiction, 2)lower boundary prediction, and 3) middle value final prediction for each given period. This can be used as a one guideline to set as bench mark model before using any other prediction models from machine/deep learning
}
\examples{
prediction_matrix= execute_roll_linear(stock,30)
}
