% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faber_3m.R
\name{faber_3m}
\alias{faber_3m}
\title{get the faber features(finance features)..}
\usage{
faber_3m(dataframe, dates_review, freq)
}
\arguments{
\item{dataframe}{multiple stock information}

\item{dates_review}{chosen period observation}

\item{frequency}{frequency forthe user chosen}
}
\value{
momentun valuebased on input dataframe stock file, dates chosen, frequency
}
\description{
get the faber features(finance features)..
}
\details{
similar to momentum 3m, except we divide by avg price of 3 months ago
it returns the momentum value based on those. Called Faber.
}
\examples{
faber_3m(stock, dates_review,freq)
}
