% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mae.R
\name{get_mae}
\alias{get_mae}
\title{get MAE score result}
\usage{
get_mae(pred, original)
}
\arguments{
\item{pred}{prediction result from any model}

\item{original}{original value at given period}
}
\value{
MAE score
}
\description{
get MAE score result
}
\details{
this will give rmse calculated score and this automatically handles the different window size of input data
}
\examples{
mae= get_mae(pred,original)
}
