% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multiple_stock.R
\name{get_multiple_stock}
\alias{get_multiple_stock}
\title{Stack each stock(what User decided to wrangle together) as a columnwise}
\usage{
get_multiple_stock(list_stocks, start, end)
}
\arguments{
\item{list}{of ticker_names}
}
\value{
multiple stacked stocks of closing price information
}
\description{
Stack each stock(what User decided to wrangle together) as a columnwise
}
\details{
this function get the <<multiple>> stock closing price informations from source, default true for getting close price from yahoo finance.
}
