% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rmse.R
\name{get_rmse}
\alias{get_rmse}
\title{get rmse score result}
\usage{
get_rmse(pred, original)
}
\arguments{
\item{pred}{prediction result from any model}

\item{original}{original value at given period}
}
\value{
rmse score
}
\description{
get rmse score result
}
\details{
this will give rmse calculated score and this automatically handles the different window size of input data
}
\examples{
rmse= get_rmse(pred,original)
}
