% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_single_stock.R
\name{get_single_stock}
\alias{get_single_stock}
\title{get the single stock from source(only closing price information)}
\usage{
get_single_stock(start, end, ticker_name, source_name)
}
\value{
closestockprice
}
\description{
get the single stock from source(only closing price information)
}
\details{
this function get the <<single>> stock information from source, default true for getting close price. You can customize the source(where to get the data)
}
\examples{
stock_1 = get_single_Stock('2001-03-12','2003-04,22','AAPL','yahoo')
}
