% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getyahooprice.R
\name{getyahooprice}
\alias{getyahooprice}
\title{get the each single stock corresponding to the ticker name from yahoo finance(OHLC information)}
\usage{
getyahooprice(tickername, start, end)
}
\value{
stock of OHLC
}
\description{
get the each single stock corresponding to the ticker name from yahoo finance(OHLC information)
}
\details{
this function get the <<single>> stock information from source, default true for getting close price from yahoo finance. Not like get_single_stock, it also gives all OHLC information of stock
}
