% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_cluster_dendogram.R
\name{give_cluster_dendogram}
\alias{give_cluster_dendogram}
\title{draw k-mean cluster analysis result in histogram table}
\usage{
give_cluster_dendogram(df_stocks)
}
\arguments{
\item{df_stocks}{bunch of stocks stacked which user want to analyze}
}
\value{
dendogram plot result from hierarchical cluster models
}
\description{
draw k-mean cluster analysis result in histogram table
}
\details{
by comparing the plot result from this function,User can analysis the stock markets regime. For instance, for each plot shows how stock movesment behaves differently. User could see the momentum of effectiveness in the market and means of the variables in each cluster. below function is another way to show the similiarities between each designated stocks User choose. Risk parity model can be combined with this. This could be another indicator of similarities of stock movement in the market at given time period
}
\examples{
draw_cluster_analysis(df_stocks, num_center, num_start)
}
