% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_summary.R
\name{give_summary}
\alias{give_summary}
\title{get summary of current market at given period}
\usage{
give_summary(stock_bunch)
}
\arguments{
\item{stock_bunch}{stacked or one matrix of multiple stock informations with only one single price information(not all OHLC)}
}
\value{
summary of principle component analysis result
}
\description{
get summary of current market at given period
}
\details{
this gives all summary of PCA analysis information result for the market
}
\examples{
give_summary(stock_bunch)
}
