% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_3m.R
\name{momentum_3m}
\alias{momentum_3m}
\title{functions related to momentum calculation}
\usage{
momentum_3m(dataframe, dates_review, freq)
}
\arguments{
\item{freq}{chosen frequency at given observation}

\item{days_end}{ending date}
}
\value{
return the momentum features with 3 months window size
}
\description{
functions related to momentum calculation
}
\details{
for each quarter we compute a momentum value
}
\examples{
momentum_3m(dataframe, dates_review,freq)
}
