% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_importance.R
\name{pca_importance}
\alias{pca_importance}
\title{get pca result importance using standard deviation and variances}
\usage{
pca_importance(x)
}
\value{
importance of varaince analysis
}
\description{
get pca result importance using standard deviation and variances
}
\details{
this function is to transform a PCA type output into a regular data frame to it can be used in Shiny as a DT output
}
\examples{
pca_importance(pca_result_vector)
}
