% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices_to_rdata.R
\name{prices_to_rdata}
\alias{prices_to_rdata}
\title{load a large csv file and will cast it as XTS data class after save this as R data format}
\usage{
prices_to_rdata(filename)
}
\value{
rdata
}
\description{
load a large csv file and will cast it as XTS data class after save this as R data format
}
\examples{
this will allow faster loading and calculation. XTS  is special time series package which handles and allow users to compute based on time frame windows.If user want to compute the return between date A and B, use do window(prices, start = A, end = B). xts class data is working for this method. dataframe, is not working since data frame sees a date as a character.
}
