% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebase.R
\name{rebase}
\alias{rebase}
\title{rescale/rebase the given input}
\usage{
rebase(vector, scale)
}
\value{
rdata
}
\description{
rescale/rebase the given input
}
\details{
comparing two price series of stocks for a given window, we need to rebase them otherwise, imagine stock 1 has a price of 100 and stock 2 a price of 1000, you cannot put on same chart, and imagine stock price grows to 150 (+50%) while stock 2 grows to 1050 (+5%). we can visualise than only if you rebase to same scale
}
\examples{
## Pass a vector and scale that you want to rescale it. rebase(vecotor,10)
}
