% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_freq.R
\name{select_freq}
\alias{select_freq}
\title{function to select quarterly data..}
\usage{
select_freq(days_end, freq)
}
\arguments{
\item{days_end}{ending date}

\item{freq}{chosen frequency at given observation}
}
\value{
return the quteraly data based on end date and frequency
}
\description{
function to select quarterly data..
}
\details{
function to select quarterly data from stored data
}
\examples{
select_freq(days_end,freq)
}
