% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdevwind.R
\name{stdevwind}
\alias{stdevwind}
\title{rolling standard deviation function}
\usage{
stdevwind(returns)
}
\value{
anualized volatility
}
\description{
rolling standard deviation function
}
\details{
this will calculated the standard deviation based on sliding window.Also, it will return the annulized standard deviation based on market live day out of 365 days in a year. calculate adjusted window frame for anualized version of sliding window based standard deviation
}
\examples{
stdevwind(stock1)
}
